quantity auto detects mask
