+++ pace-util/pace/util/grid/__init__.py
@@ -15,6 +15,7 @@ from .helper import (
     AngleGridData,
     ContravariantGridData,
     DampingCoefficients,
+    DriverGridData,
     GridData,
     HorizontalGridData,
     VerticalGridData,
