+++ pace-util/pace/util/grid/generation.py
@@ -63,7 +63,6 @@ class MetricTerms:
     CARTESIAN_DIM = "xyz_direction"
     N_TILES = 6
     RIGHT_HAND_GRID = False
-    UNIT_X_DIM = "unit_vector_x"
 
     def __init__(
         self,
@@ -85,7 +84,7 @@ class MetricTerms:
                 self.LON_OR_LAT_DIM: 2,
                 self.TILE_DIM: 6,
                 self.CARTESIAN_DIM: 3,
-                self.UNIT_X_DIM: 1,
+                fv3util.X_DIM: 1,
             }
         )
         self._grid_indexing = GridIndexing.from_sizer_and_communicator(
@@ -2168,10 +2167,10 @@ class MetricTerms:
         edge_s = self.quantity_factory.zeros([fv3util.X_INTERFACE_DIM], "")
         edge_n = self.quantity_factory.zeros([fv3util.X_INTERFACE_DIM], "")
         edge_e = self.quantity_factory.zeros(
-            [self.UNIT_X_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
         )
         edge_w = self.quantity_factory.zeros(
-            [self.UNIT_X_DIM, fv3util.Y_INTERFACE_DIM], ""
+            [fv3util.X_DIM, fv3util.Y_INTERFACE_DIM], ""
         )
         (
             edge_w.data[:, nhalo:-nhalo],
