+++ pace-util/pace/util/grid/helper.py
@@ -1,8 +1,9 @@
 import dataclasses
-from typing import Any, Optional, Sequence
+from typing import Any, Optional
 
-import pace.util
-from pace.dsl.stencil import GridIndexing
+import numpy as np
+
+from pace.dsl import gt4py_utils as utils
 from pace.dsl.typing import FloatFieldI, FloatFieldIJ
 
 from .generation import MetricTerms
@@ -472,12 +473,107 @@ class GridData:
         return self._angle_data.cos_sg4
 
 
-def quantity_wrap(storage, dims: Sequence[str], grid_indexing: GridIndexing):
-    origin, extent = grid_indexing.get_origin_domain(dims)
-    return pace.util.Quantity(
-        storage,
-        dims=dims,
-        units="unknown",
-        origin=origin,
-        extent=extent,
-    )
+@dataclasses.dataclass(frozen=True)
+class DriverGridData:
+    """
+    Terms used to Apply Physics changes to the Dycore.
+    Attributes:
+      vlon1: x-component of unit lon vector in eastward longitude direction
+      vlon2: y-component of unit lon vector in eastward longitude direction
+      vlon3: z-component of unit lon vector in eastward longitude direction
+      vlat1: x-component of unit lat vector in northward latitude direction
+      vlat2: y-component of unit lat vector in northward latitude direction
+      vlat3: z-component of unit lat vector in northward latitude direction
+      edge_vect_w: factor to interpolate A to C grids at the western grid edge
+      edge_vect_e: factor to interpolate A to C grids at the easter grid edge
+      edge_vect_s: factor to interpolate A to C grids at the southern grid edge
+      edge_vect_n: factor to interpolate A to C grids at the northern grid edge
+      es1_1: x-component of grid local unit vector in x-direction at cell edge
+      es1_2: y-component of grid local unit vector in x-direction at cell edge
+      es1_3: z-component of grid local unit vector in x-direction at cell edge
+      ew2_1: x-component of grid local unit vector in y-direction at cell edge
+      ew2_2: y-component of grid local unit vector in y-direction at cell edge
+      ew2_3: z-component of grid local unit vector in y-direction at cell edge
+    """
+
+    vlon1: FloatFieldIJ
+    vlon2: FloatFieldIJ
+    vlon3: FloatFieldIJ
+    vlat1: FloatFieldIJ
+    vlat2: FloatFieldIJ
+    vlat3: FloatFieldIJ
+    edge_vect_w: FloatFieldI
+    edge_vect_e: FloatFieldI
+    edge_vect_s: FloatFieldIJ
+    edge_vect_n: FloatFieldIJ
+    es1_1: FloatFieldIJ
+    es1_2: FloatFieldIJ
+    es1_3: FloatFieldIJ
+    ew2_1: FloatFieldIJ
+    ew2_2: FloatFieldIJ
+    ew2_3: FloatFieldIJ
+
+    @classmethod
+    def new_from_metric_terms(cls, metric_terms: MetricTerms) -> "DriverGridData":
+        # TODO fix <Quantity>.storage mask for FieldI
+        shape = metric_terms.lon.data.shape
+        backend = metric_terms.edge_vect_n.gt4py_backend
+        edge_vect_n = utils.make_storage_data(
+            metric_terms.edge_vect_n.data,
+            (shape[0],),
+            axis=0,
+            backend=backend,
+        )
+        edge_vect_s = utils.make_storage_data(
+            metric_terms.edge_vect_s.data,
+            (shape[0],),
+            axis=0,
+            backend=backend,
+        )
+        east_edge_data = metric_terms.edge_vect_e.data[np.newaxis, ...]
+        east_edge_data = np.repeat(east_edge_data, shape[0], axis=0)
+        edge_vect_e = utils.make_storage_data(
+            data=east_edge_data,
+            shape=east_edge_data.shape,
+            origin=(0, 0),
+            backend=backend,
+        )
+        west_edge_data = metric_terms.edge_vect_w.data[np.newaxis, ...]
+        west_edge_data = np.repeat(west_edge_data, shape[0], axis=0)
+        edge_vect_w = utils.make_storage_data(
+            data=west_edge_data,
+            shape=west_edge_data.shape,
+            origin=(0, 0),
+            backend=backend,
+        )
+
+        vlon1, vlon2, vlon3 = metric_terms.split_cartesian_into_storages(
+            metric_terms.vlon
+        )
+        vlat1, vlat2, vlat3 = metric_terms.split_cartesian_into_storages(
+            metric_terms.vlat
+        )
+        es1_1, es1_2, es1_3 = metric_terms.split_cartesian_into_storages(
+            metric_terms.es1
+        )
+        ew2_1, ew2_2, ew2_3 = metric_terms.split_cartesian_into_storages(
+            metric_terms.ew2
+        )
+        return cls(
+            vlon1=vlon1,
+            vlon2=vlon2,
+            vlon3=vlon3,
+            vlat1=vlat1,
+            vlat2=vlat2,
+            vlat3=vlat3,
+            es1_1=es1_1,
+            es1_2=es1_2,
+            es1_3=es1_3,
+            ew2_1=ew2_1,
+            ew2_2=ew2_2,
+            ew2_3=ew2_3,
+            edge_vect_w=edge_vect_w,
+            edge_vect_e=edge_vect_e,
+            edge_vect_s=edge_vect_s,
+            edge_vect_n=edge_vect_n,
+        )
