+++ stencils/pace/stencils/update_atmos_state.py
@@ -5,8 +5,7 @@ from fv3gfs.physics.physics_state import PhysicsState
 from pace.dsl.stencil import StencilFactory
 from pace.dsl.typing import Float, FloatField
 from pace.stencils.fv_update_phys import ApplyPhysics2Dycore
-from pace.stencils.testing.grid import DriverGridData
-from pace.util.grid import GridData
+from pace.util.grid import DriverGridData, GridData
 
 
 def fill_gfs(pe: FloatField, q: FloatField, q_min: Float):
